+++ test/model.test.js
@@ -3826,6 +3826,47 @@ describe('Model', function() {
         });
       });
     });
+    it('is saved object with proper defaults', function() {
+      const schema = new Schema({
+        foo: {
+          x: { type: String },
+          y: { type: String }
+        },
+        boo: {
+          x: { type: Boolean, default: false }
+        },
+        bee: {
+          x: { type: Boolean, default: false },
+          y: { type: Boolean, default: false }
+        }
+      });
+      const Test = db.model('Test', schema);
+
+      const doc = new Test({
+        foo: { x: 'a', y: 'b' },
+        bee: {},
+        boo: {}
+      });
+
+      return co(function*() {
+        yield doc.save();
+        assert.equal(doc.bee.x, false);
+        assert.equal(doc.bee.y, false);
+        assert.equal(doc.boo.x, false);
+
+        doc.bee = undefined;
+        doc.boo = undefined;
+
+        yield doc.save();
+
+        const docAfterUnsetting = yield Test.findById(doc._id);
+
+        assert.equal(docAfterUnsetting.bee.x, false);
+        assert.equal(docAfterUnsetting.bee.y, false);
+        assert.equal(docAfterUnsetting.boo.x, false);
+      });
+    });
+
   });
 
   it('path is cast to correct value when retreived from db', function(done) {
@@ -4079,8 +4120,7 @@ describe('Model', function() {
 
       return co(function*() {
         yield Location.collection.drop().catch(() => {});
-        yield Location.createCollection();
-        yield Location.createIndexes();
+        yield Location.init();
 
         yield Location.create({
           name: 'Undefined location'
@@ -5194,7 +5234,7 @@ describe('Model', function() {
             const db = yield start();
             const MyModel = db.model('Test', new Schema({ name: String }));
 
-            yield MyModel.createCollection();
+            yield MyModel.init();
 
             const changeStream = MyModel.watch();
             const closed = new global.Promise(resolve => {
@@ -7323,7 +7363,8 @@ describe('Model', function() {
 
         yield User.insertMany([
           new User({ name: 'Hafez1_gh-9673-2' }),
-          new User({ name: 'Hafez2_gh-9673-2' })
+          new User({ name: 'Hafez2_gh-9673-2' }),
+          new User({ name: 'Hafez3_gh-9673-2' })
         ]);
 
         const users = yield User.find().sort('name');
@@ -7339,7 +7380,8 @@ describe('Model', function() {
           usersAfterUpdate.map(user => user.name),
           [
             'Hafez1_gh-9673-2-updated',
-            'Hafez2_gh-9673-2-updated'
+            'Hafez2_gh-9673-2-updated',
+            'Hafez3_gh-9673-2'
           ]
         );
       });
@@ -7412,7 +7454,7 @@ describe('Model', function() {
         assert.equal(user2.isNew, false);
       });
     });
-    it('sets `isNew` to false when a document succeds and `isNew` does not change when some fail', () => {
+    it('sets `isNew` to false when a document succeeds and `isNew` does not change when some fail', () => {
       return co(function* () {
         const userSchema = new Schema({
           name: { type: String, unique: true }
@@ -7502,6 +7544,20 @@ describe('Model', function() {
         const usersFromDatabase = yield User.find({ _id: { $in: [user1._id, user2._id] } }).sort('_id');
         assert.equal(usersFromDatabase[0].name, 'name from pre-save');
         assert.equal(usersFromDatabase[1].name, 'name from pre-save');
+      });
+    });
+    it('works if some document is not modified (gh-10437)', () => {
+      const userSchema = new Schema({
+        name: String
+      });
+
+      const User = db.model('User', userSchema);
+
+      return co(function*() {
+        const user = yield User.create({ name: 'Hafez' });
+
+        const err = yield User.bulkSave([user]).then(() => null, err => err);
+        assert.ok(err == null);
       });
     });
   });